multi: move node ann validation cod to netnn
